simplify constrained rust type example
