+++ interfaces/kalosm/examples/constrained-rust-types.rs
@@ -7,9 +7,9 @@ struct Account {
     /// A summary of the account holder
     #[parse(pattern = r"[a-zA-Z,.?!\d ]{1,80}")]
     summary: String,
-    /// The user name of the account holder. This may be the full name of the user or a pseudonym
+    /// The name of the account holder. This may be the full name of the user or a pseudonym
     #[parse(pattern = "[a-zA-Z ]{1,20}")]
-    username: String,
+    name: String,
     /// The age of the account holder
     #[parse(range = 1..=100)]
     age: u8,
@@ -21,33 +21,18 @@ async fn main() -> anyhow::Result<()> {
     let llm = Llama::new_chat().await?;
 
     // Create a task that generates a list of accounts
-    let task = llm.task(
-        "You generate accounts based on a description of the account holder",
-    )
-    .with_example(
-        "John Doe is a 25 year old accountant who works at ABC Inc.",
-        r#"{ "summary": "John Doe is 25", "username": "John Doe", "age": 25 }"#,
-    )
-    .with_example(
-        "A 57 years old woman named Candace.",
-        r#"{ "summary": "Candace is 57", "username": "Candace", "age": 57 }"#,
-    )
-    .with_example(
-        "Anderson is a programmer who works at Google, but he goes by Mr Robot. He is 71 years old.",
-        r#"{ "summary": "Mr Robot is 71", "username": "Mr Robot", "age": 71 }"#,
-    );
+    let task = llm
+        .task("You generate accounts based on a description of the account holder")
+        // The typed combinator constraints the task for the default parser for the type you get out. (In
+        // this case, Account)
+        .typed();
 
-    loop {
-        let prompt = prompt_input("Describe an account holder: ").unwrap();
-        let mut stream = task(&prompt).typed();
+    // Task can be called like a function with the input to the task. You can await the stream, modify the
+    // constraints, or sampler
+    let account: Account =
+        task("Candice is the CEO of a fortune 500 company. She is a 30 years old.").await?;
 
-        // Stream the output to the console
-        stream.to_std_out().await?;
-
-        // And get the typed result once the stream is finished
-        let accounts: Account = stream.await?;
-        println!("\n{:#?}", accounts);
-    }
+    println!("{:#?}", account);
 
     Ok(())
 }
